fix(ble): prevent atomic counter underflow on disconnect/unsubscribe
CONNECTED and SUBSCRIBED used fetch_sub which wraps to u32::MAX on
underflow (overflow-checks = false in release). Replace with
fetch_update + saturating_sub so spurious callbacks from NimBLE never
corrupt the counters.

Also log BLE advertising errors instead of silently discarding them,
remove a POLL_MS constant that shadowed the imported buttons::POLL_MS,
and normalize slot numbers to u8 throughout (cli, fingerprint, display)
to eliminate scattered as-casts at the application level.

+++ firmware/src/display.rs
@@ -201,7 +201,7 @@ pub fn show_status_2line<D: DrawTarget<Color = Rgb565>>(
 }
 
 /// Successful enrollment.
-pub fn show_enroll_ok<D: DrawTarget<Color = Rgb565>>(d: &mut D, sb: &StatusBar<'_>, slot: u16) {
+pub fn show_enroll_ok<D: DrawTarget<Color = Rgb565>>(d: &mut D, sb: &StatusBar<'_>, slot: u8) {
     clear_content(d);
     update_topbar(d, sb);
     draw_center(d, "ENROLLED", CONTENT_CY - 20, NEON_GREEN);
